start on expr caching
